+++ src/components/Nav/AuthNav.jsx
@@ -1,9 +1,5 @@
-import React from "react";
-import { DarkTheme } from "./DarkTheme";
-import { Logo } from "./Logo";
+import { DarkTheme, Logo, PublicRoutesNav, SearchBar } from ".";
 import "./nav.css";
-import { PublicRoutesNav } from "./PublicRoutesNav";
-import { SearchBar } from "./SearchBar";
 export const AuthNav = () => {
   return (
     <header className="bg-primary text text-sm nav-header center-xy">
