+++ src/components/Nav/DarkTheme.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useTheme } from "../../context/theme-context";
+import { useTheme } from "context";
 
 export const DarkTheme = () => {
   const { theme, toggleDarkTheme } = useTheme();
