+++ src/components/Nav/index.js
@@ -1,0 +1,8 @@
+export { AuthNav } from "./AuthNav";
+export { DarkTheme } from "./DarkTheme";
+export { Logo } from "./Logo";
+export { MainNav } from "./MainNav";
+export { PrivateRoutesNav } from "./PrivateRoutesNav";
+export { PublicRoutesNav } from "./PublicRoutesNav";
+export { SearchBar } from "./SearchBar";
+export { SiteAuthNav } from "./SiteAuthNav";
