+++ src/components/RequiresAuth/index.jsx
@@ -1,5 +1,5 @@
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuth } from "../../context/auth-context"
+import { useAuth } from "context"
 export function RequiresAuth({ children }) {
   const location = useLocation();
   const { isLoggedIn } = useAuth();
