+++ src/components/SortAndFilter/Rating.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useProductListing } from "../../context/product-listing-context";
+import { useProductListing } from "context";
 
 export function Rating() {
   const { productsListingState, setRating } = useProductListing();
