+++ src/components/index.js
@@ -1,0 +1,8 @@
+export { Alert } from "./Alert";
+export { Card } from "./Card";
+export { Footer } from "./Footer";
+export { ImageSlider } from "./ImageSlider";
+export { AuthNav, MainNav } from "./Nav";
+export { RequiresAuth } from "./RequiresAuth";
+export { SortAndFilter } from "./SortAndFilter";
+export { AuthLayout, MainLayout } from "./layouts";
