+++ src/context/index.js
@@ -1,0 +1,9 @@
+export { AlertProvider, useAlert } from "./alert-context";
+export { AuthProvider, useAuth } from "./auth-context";
+export { useCart, CartProvider } from "./cart-context";
+export { CategoriesProvider, useCategories } from "./categories-context";
+export { ProductsListingProvider, useProductListing } from "./product-listing-context";
+export {  ThemeProvider, useTheme  } from "./theme-context";
+export { UserProvider, useUser } from "./user-context";
+export { WindowDimensionProvider, useWindowDimension } from "./window-context";
+export { WishlistProvider, useWishlist } from "./wishlist-context";
