+++ src/index.js
@@ -4,15 +4,8 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { ProductsListingProvider } from "./context/product-listing-context";
-import { CategoriesProvider } from "./context/categories-context";
-import { WindowDimensionProvider } from "./context/window-context";
-import { AuthProvider } from "./context/auth-context";
-import { UserProvider } from "./context/user-context";
-import { AlertProvider } from "./context/alert-context";
-import { CartProvider } from "./context/cart-context";
-import { WishlistProvider } from "./context/wishlist-context";
-import { ThemeProvider } from "./context/theme-context";
+import { AlertProvider, AuthProvider, CartProvider, CategoriesProvider, ProductsListingProvider, ThemeProvider, UserProvider, WindowDimensionProvider, WishlistProvider } from "context";
+
 // Call make Server
 makeServer();
