+++ src/pages/index.js
@@ -3,4 +3,4 @@ export { Wishlist } from "./wishlist";
 export { Login, SignUp, ForgotPassword } from "./auth";
 export { Cart } from "./cart";
 export { ProductListing } from "./products";
-export { NotFound } from "./notFound/NotFound";
+export { NotFound } from "./notFound";
