+++ src/reducers/index.js
@@ -1,0 +1,2 @@
+export { categoriesReducer } from "./categoriesReducer";
+export { productsListingReducer } from "./productsListingReducer";
