+++ src/services/index.js
@@ -1,0 +1,3 @@
+export { addItem } from "./addItem";
+export { removeItem } from "./removeItem";
+export { updateCartItem } from "./updateCartItem";
