+++ src/utils/index.js
@@ -1,0 +1,5 @@
+export { checkActive } from "./checkActive";
+export { doArraysHaveCommonElement } from "./doArraysHaveCommonElement";
+export { numSort } from "./numSort";
+export { slugify } from "./slugify";
+export { subsetOfArray } from "./subsetOfArray";
